+++ ChangeLog.md
@@ -6,6 +6,8 @@
     * Moved setup of Room, Portal, Box, Sector, ... into their
       Constructors.
     * Fixed menu play button
+    * Removed Render flags fSprite, fRoomModel, fPortals
+    * Room BoundingBox now using Box object
 
     [ 20140504 ]
     * Forcing use of new Room and Sprite classes in World.